refactor: reorganize widget headers
