+++ include/imguix/widgets/time/date_picker.hpp
@@ -11,7 +11,7 @@
 #include <cstdint>
 #include <cstdio>
 #include <string>
-#include "arrow_stepper.hpp"
+#include <imguix/widgets/input/arrow_stepper.hpp>
 #include <imguix/utils/time_utils.hpp>  // days_from_civil, civil_from_days, num_days_in_month, clamp_ymdhms, month_short_name
 #include <imguix/extensions/sizing.hpp> // ImGuiX::Extensions::CalcDateComboWidth(), etc.
